chore: wip n-api
