+++ src/packages/engine-core/src/NodeEngine.ts
@@ -1,9 +1,5 @@
 import { getEnginesPath } from '@prisma/engines'
-import {
-  ConnectorType,
-  DataSource,
-  GeneratorConfig,
-} from '@prisma/generator-helper'
+import { ConnectorType, GeneratorConfig } from '@prisma/generator-helper'
 import { getPlatform, Platform } from '@prisma/get-platform'
 import chalk from 'chalk'
 import { ChildProcessByStdio, spawn } from 'child_process'
@@ -26,7 +22,7 @@ import {
   PrismaClientRustPanicError,
   PrismaClientUnknownRequestError,
   RequestError,
-} from './Engine'
+} from './errors'
 import {
   convertLog,
   isRustError,
@@ -39,46 +35,19 @@ import { omit } from './omit'
 import { printGeneratorConfig } from './printGeneratorConfig'
 import { Undici } from './undici'
 import { fixBinaryTargets, getRandomString, plusX } from './util'
+import {
+  DatasourceOverwrite,
+  Engine,
+  EngineConfig,
+  GetConfigResult,
+} from './Engine'
 
 const debug = Debug('prisma:engine')
 const exists = promisify(fs.exists)
 
-export interface DatasourceOverwrite {
-  name: string
-  url: string
-}
-
 // eslint-disable-next-line
 const logger = (...args) => {
   // console.log(chalk.red.bold('logger '), ...args)
-}
-
-export interface EngineConfig {
-  cwd?: string
-  dirname?: string
-  datamodelPath: string
-  enableDebugLogs?: boolean
-  enableEngineDebugMode?: boolean // dangerous! https://github.com/prisma/prisma-engines/issues/764
-  prismaPath?: string
-  fetcher?: (query: string) => Promise<{ data?: any; error?: any }>
-  generator?: GeneratorConfig
-  datasources?: DatasourceOverwrite[]
-  showColors?: boolean
-  logQueries?: boolean
-  logLevel?: 'info' | 'warn'
-  env?: Record<string, string>
-  flags?: string[]
-  useUds?: boolean
-
-  clientVersion?: string
-  enableExperimental?: string[]
-  engineEndpoint?: string
-  activeProvider?: string
-}
-
-type GetConfigResult = {
-  datasources: DataSource[]
-  generators: GeneratorConfig[]
 }
 
 /**
@@ -113,7 +82,6 @@ export type StopDeferred = {
   resolve: (code: number | null) => void
   reject: (err: Error) => void
 }
-export type EngineEventType = 'query' | 'info' | 'warn' | 'error' | 'beforeExit'
 
 const engines: NodeEngine[] = []
 const socketPaths: string[] = []
@@ -121,7 +89,7 @@ const socketPaths: string[] = []
 const MAX_STARTS = process.env.PRISMA_CLIENT_NO_RETRY ? 1 : 2
 const MAX_REQUEST_RETRIES = process.env.PRISMA_CLIENT_NO_RETRY ? 1 : 2
 
-export class NodeEngine {
+export class NodeEngine implements Engine {
   private logEmitter: EventEmitter
   private showColors: boolean
   private logQueries: boolean
@@ -348,7 +316,7 @@ You may have to run ${chalk.greenBright(
     }
   }
 
-  async getPlatform(): Promise<Platform> {
+  private async getPlatform(): Promise<Platform> {
     // eslint-disable-next-line @typescript-eslint/no-misused-promises
     if (this.platformPromise) {
       return this.platformPromise
@@ -532,7 +500,7 @@ ${chalk.dim("In case we're mistaken, please report this to us 🙏.")}`)
     return printGeneratorConfig(fixedGenerator)
   }
 
-  printDatasources(): string {
+  private printDatasources(): string {
     if (this.datasources) {
       return JSON.stringify(this.datasources)
     }
@@ -920,7 +888,7 @@ You very likely have the wrong "binaryTarget" defined in the schema.prisma file.
   /**
    * Use the port 0 trick to get a new port
    */
-  protected getFreePort(): Promise<number> {
+  private getFreePort(): Promise<number> {
     return new Promise((resolve, reject) => {
       const server = net.createServer((s) => s.end(''))
       server.unref()
@@ -948,7 +916,7 @@ You very likely have the wrong "binaryTarget" defined in the schema.prisma file.
     return this.getConfigPromise
   }
 
-  async _getConfig(): Promise<GetConfigResult> {
+  private async _getConfig(): Promise<GetConfigResult> {
     const prismaPath = await this.getPrismaPath()
 
     const env = await this.getEngineEnvVars()
